Edit vuex logics

+++ src/store/directories/actions.js
@@ -1,19 +1,20 @@
-export async function CreateDirectory ({
-  state,
-  commit,
-  getters
-}, title, id) {
+export async function CreateDirectory (store, title, id) {
   let res = await fetch(`https://challenge.spexa.dev/directory/${id}`, {
     method: 'POST',
     mode: 'cors',
     headers: {
-      Authorization: `Bearer ${getters['auth/AccessToken']}`
+      Authorization: `Bearer ${localStorage.getItem('token') || store.rootState.auth.token}`
     },
     body: JSON.stringify({ title })
   })
+  const response = await res.json()
+  if (res.status === 401 && response.message.include('JWT')) {
+    // when our response return 401 status or refreshToken will refresh token and callback our function again
+    store.dispatch('auth/refreshToken', await deleteDirectory(store, title, id), { root: true })
+  }
   try {
     res = await res.json()
-    commit('setter', ['directory', res.data.directories])
+    store.commit('setter', ['directory', res.data.directories])
     return res
   } catch (e) {
     return res.message
@@ -31,13 +32,14 @@ export async function getDirectory (store, id) {
     })
   })
   // handling 401 error
-  if (res.status === 401) {
+  const response = await res.json()
+  if (res.status === 401 && response.message.include('JWT')) {
     // when our response return 401 status or refreshToken will refresh token and callback our function again
     store.dispatch('auth/refreshToken', await getDirectory(store, id), { root: true })
   }
 
   try {
-    res = await res.json()
+    res = response
     store.commit('setter', ['directory', res.data.directories])
     return res
   } catch (e) {
@@ -57,14 +59,14 @@ export async function deleteDirectory (store, title, id) {
     })
 
   })
-
-  if (res.status === 401) {
+  const response = await res.json()
+  if (res.status === 401 && response.message.include('JWT')) {
     // when our response return 401 status or refreshToken will refresh token and callback our function again
     store.dispatch('auth/refreshToken', await deleteDirectory(store, title, id), { root: true })
   }
 
   try {
-    res = await res.json()
+    res = response
     return res.data
   } catch {
     return res.message
